Changed:
-Wait and aquire swapchain image code now moved to EndFrame(), allowing the CPU to complete a frame before waiting. Then when the wait happens it won't be as long.
-Buffer Objects are now allocated using CLib::Allocator.

Fixed:

-Use of PB_ASSERT for assertions with messages.
-Lots of warnings

+++ VKR/VKR/FramebufferCache.cpp
@@ -15,8 +15,8 @@ namespace PB
 
 	size_t FramebufferDescHasher::operator()(const FramebufferDesc& desc) const
 	{
-		auto descHash = MurmurHash3_x64_64(&desc, sizeof(FramebufferDesc) - sizeof(FramebufferDesc::m_attachmentViews), 0);
-		return MurmurHash3_x64_64(desc.m_attachmentViews, sizeof(TextureView) * desc.m_attachmentCount, descHash);
+		auto descHash = MurmurHash3_x64_64(&desc, static_cast<int>(sizeof(FramebufferDesc) - sizeof(FramebufferDesc::m_attachmentViews)), 0);
+		return MurmurHash3_x64_64(desc.m_attachmentViews, static_cast<int>(sizeof(TextureView) * desc.m_attachmentCount), descHash);
 	}
 
 	void FramebufferCache::Init(Device* device)
